Create README.md
This notebook is trying to build a model which will predict Bollywood celebrity the given image looks like. This project is trained on 100 Bollywood celebrities and the dataset is taken from kaggle. The dataset is taken from Kaggle: Bollywood celeb localized face dataset You can follow the analysis on Kaggle